requirements.txt
django==6.0.4
djangorestframework==3.17.1
pypdf==6.10.2
